+++ InscryptionCommunityPatch/PixelTutor/PixelPlayableCardArray.cs
@@ -1,7 +1,9 @@
 using DiskCardGame;
 using GBC;
 using InscryptionAPI;
+using InscryptionAPI.Helpers;
 using InscryptionAPI.Helpers.Extensions;
+using InscryptionCommunityPatch.Card;
 using Pixelplacement;
 using Sirenix.Utilities;
 using System.Collections;
@@ -16,7 +18,10 @@ public class PixelPlayableCardArray : ManagedBehaviour
     public readonly List<PixelPlayableCard> displayedCards = new();
 
     private PixelPlayableCard gameObjectReference = null;
+    private GenericUIButton buttonReference = null;
     private GameObject overlay;
+    private GenericUIButton forwardButton = null;
+    private GenericUIButton backButton = null;
 
     private readonly int maxRows = 6;
     private readonly int maxCardsPerRow = 7;
@@ -27,24 +32,39 @@ public class PixelPlayableCardArray : ManagedBehaviour
     private readonly float tweenDurationModifier = 0.5f;
     private Vector3 offscreenPositionOffset = new(2f, 2f, -4f);
 
+    private int numPages = 0;
+    private int currentPageIndex = 0;
+
     private GamepadGridControl gamepadGrid;
 
     public IEnumerator DisplayUntilCancelled(List<CardInfo> cards,
         Func<bool> cancelCondition, Action cardsPlacedCallback = null)
     {
+        currentPageIndex = 0;
+        numPages = 1 + (cards.Count - 1) / (maxCardsPerRow * maxRows);
         InitializeGamepadGrid();
-        yield return SpawnAndPlaceCards(cards, GetNumRows(cards.Count), isDeckReview: true);
+        yield return SpawnAndPlaceCards(cards, GetNumRows(cards.Count), 0, isDeckReview: true);
+        InitialiseButtons(cards, true, true);
         cardsPlacedCallback?.Invoke();
         SetCardsEnabled(enabled: true);
+        if (numPages > 1)
+        {
+            forwardButton.gameObject.SetActive(true);
+            backButton.gameObject.SetActive(true);
+            EnableButtons(true);
+        }
         yield return new WaitUntil(() => cancelCondition());
         SetCardsEnabled(enabled: false);
         yield return CleanUpCards();
+        CleanUpButtons();
     }
 
     public IEnumerator SelectPixelCardFrom(
         List<CardInfo> cards, Action<PixelPlayableCard> cardSelectedCallback,
         Func<bool> cancelCondition = null, bool forPositiveEffect = true)
     {
+        currentPageIndex = 0;
+        numPages = 1 + (cards.Count - 1) / (maxCardsPerRow * maxRows);
         if (cards.Count == 0)
             yield break;
 
@@ -53,13 +73,14 @@ public class PixelPlayableCardArray : ManagedBehaviour
             PixelPlayableCard[] objs = Resources.FindObjectsOfTypeAll<PixelPlayableCard>();
             objs.Sort((a, b) => a.Anim.cardRenderer.sortingGroupOrder - b.Anim.cardRenderer.sortingGroupOrder);
             gameObjectReference = Instantiate(objs[0], base.transform);
+            gameObjectReference.name = "TutorCardObjRef";
             gameObjectReference.transform.localPosition = new(-100f, -100f, 0); // go away
             gameObjectReference.transform.rotation.Set(0, 0, 0, 0);
         }
 
         if (overlay == null)
         {
-            GameObject fade = Resources.FindObjectsOfTypeAll<PauseMenu>()[0].menuParent.FindChild("GameFade");
+            GameObject fade = PauseMenu.instance.menuParent.FindChild("GameFade");
             overlay = Instantiate(fade, base.transform);
             overlay.name = "TutorUIFade";
             var renderer = overlay.GetComponent<SpriteRenderer>();
@@ -70,9 +91,19 @@ public class PixelPlayableCardArray : ManagedBehaviour
         }
 
         InitializeGamepadGrid();
-        yield return SpawnAndPlaceCards(cards, GetNumRows(cards.Count), isDeckReview: false, forPositiveEffect);
+        
+        yield return SpawnAndPlaceCards(cards, GetNumRows(cards.Count), 0, false, forPositiveEffect);
+        InitialiseButtons(cards, false, forPositiveEffect);
+
         yield return new WaitForSeconds(0.15f);
         SetCardsEnabled(enabled: true);
+        if (numPages > 1)
+        {
+            forwardButton.gameObject.SetActive(true);
+            backButton.gameObject.SetActive(true);
+            EnableButtons(true);
+        }
+
         selectedCard = null;
         yield return new WaitUntil(() => selectedCard != null || (cancelCondition != null && cancelCondition()));
         SetCardsEnabled(enabled: false);
@@ -80,18 +111,87 @@ public class PixelPlayableCardArray : ManagedBehaviour
             displayedCards.Remove(selectedCard);
 
         yield return CleanUpCards();
+        CleanUpButtons();
+        forwardButton = backButton = null;
+
         if (selectedCard != null)
         {
             cards.Remove(selectedCard.Info);
             cardSelectedCallback?.Invoke(selectedCard);
         }
     }
+    private void InitialiseButtons(List<CardInfo> cards, bool deckReview, bool forPositiveEffect)
+    {
+        if (forwardButton == null)
+        {
+            forwardButton = Instantiate(Singleton<HammerButton>.Instance.button, base.transform);
+            forwardButton.name = "TutorForwardButton";
+            forwardButton.cursorType = CursorType.MoveUp;
+            forwardButton.defaultSprite = TextureHelper.GetImageAsSprite("forwardButton.png", typeof(PixelPlayableCardArray).Assembly, TextureHelper.SpriteType.PixelStandardButton);
+            forwardButton.hoveringSprite = TextureHelper.GetImageAsSprite("forwardButton_hover.png", typeof(PixelPlayableCardArray).Assembly, TextureHelper.SpriteType.PixelStandardButton);
+            forwardButton.disabledSprite = TextureHelper.GetImageAsSprite("forwardButton_disabled.png", typeof(PixelPlayableCardArray).Assembly, TextureHelper.SpriteType.PixelStandardButton);
+            forwardButton.downSprite = TextureHelper.GetImageAsSprite("forwardButton_down.png", typeof(PixelPlayableCardArray).Assembly, TextureHelper.SpriteType.PixelStandardButton);
+            forwardButton.transform.position = new(1.25f, -0.75f, 0f);
+            forwardButton.CursorSelectStarted += delegate (MainInputInteractable x)
+            {
+                OnChangePage(true, cards, deckReview, forPositiveEffect);
+            };
 
+            backButton = Instantiate(Singleton<HammerButton>.Instance.button, base.transform);
+            backButton.name = "TutorBackButton";
+            backButton.cursorType = CursorType.MoveDown;
+            backButton.defaultSprite = TextureHelper.GetImageAsSprite("backButton.png", typeof(PixelPlayableCardArray).Assembly, TextureHelper.SpriteType.PixelStandardButton);
+            backButton.hoveringSprite = TextureHelper.GetImageAsSprite("backButton_hover.png", typeof(PixelPlayableCardArray).Assembly, TextureHelper.SpriteType.PixelStandardButton);
+            backButton.disabledSprite = TextureHelper.GetImageAsSprite("backButton_disabled.png", typeof(PixelPlayableCardArray).Assembly, TextureHelper.SpriteType.PixelStandardButton);
+            backButton.downSprite = TextureHelper.GetImageAsSprite("backButton_down.png", typeof(PixelPlayableCardArray).Assembly, TextureHelper.SpriteType.PixelStandardButton);
+            backButton.transform.position = new(1.25f, -1f, 0f);
+            backButton.CursorSelectStarted += delegate (MainInputInteractable x)
+            {
+                OnChangePage(false, cards, deckReview, forPositiveEffect);
+            };
+        }
+        forwardButton.gameObject.SetActive(false);
+        backButton.gameObject.SetActive(false);
+        EnableButtons(false);
+    }
+    private void CleanUpButtons()
+    {
+        if (forwardButton.gameObject.active)
+        {
+            forwardButton.gameObject.SetActive(false);
+            backButton.gameObject.SetActive(false);
+        }
+        //Destroy(forwardButton.gameObject);
+        //Destroy(backButton.gameObject);
+    }
+    private void EnableButtons(bool enable)
+    {
+        forwardButton.SetEnabled(enable);
+        backButton.SetEnabled(enable);
+    }
+    private void OnChangePage(bool forwards, List<CardInfo> cards, bool deckReview, bool forPositiveEffect)
+    {
+        EnableButtons(false);
+        SetCardsEnabled(enabled: false);
+        base.StartCoroutine(ChangePage(forwards, cards, deckReview, forPositiveEffect));
+    }
+    private IEnumerator ChangePage(bool forwards, List<CardInfo> cards, bool deckReview, bool forPositiveEffect)
+    {
+        yield return CleanUpCards();
+
+        if (forwards)
+            currentPageIndex = (currentPageIndex + 1) >= numPages ? 0 : currentPageIndex + 1;
+        else
+            currentPageIndex = (currentPageIndex - 1) < 0 ? (numPages - 1) : currentPageIndex - 1;
+
+        yield return SpawnAndPlaceCards(cards, GetNumRows(cards.Count), currentPageIndex, deckReview, forPositiveEffect);
+        SetCardsEnabled(enabled: true);
+        EnableButtons(true);
+    }
     protected Vector3 GetCardPosition(int cardIndex, int cardsToPlace, int numRows, int cardsPerRow)
     {
         int cardRowIndex = Mathf.FloorToInt(cardIndex / cardsPerRow);
         float xPos = GetXPos(cardIndex, cardRowIndex, cardsToPlace, cardsPerRow); // [-1.75, 0.7]
-
         float yPos = Mathf.Max(-0.801f, 0.451f - ((8.5f - numRows) * 0.1f * cardRowIndex)); // [0.451, -0.801]
         float zPos = 0f + 0.01f * cardRowIndex;
         return new Vector3(xPos, yPos, zPos);
@@ -131,7 +231,7 @@ public class PixelPlayableCardArray : ManagedBehaviour
         SetBoardEnabled(true);
     }
 
-    protected IEnumerator SpawnAndPlaceCards(List<CardInfo> cards, int numRows, bool isDeckReview = false, bool forPositiveEffect = true)
+    protected IEnumerator SpawnAndPlaceCards(List<CardInfo> cards, int numRows, int pageIndex, bool isDeckReview = false, bool forPositiveEffect = true)
     {
         SetOverlayEnabled(true);
         SetBoardEnabled(false);
@@ -141,11 +241,17 @@ public class PixelPlayableCardArray : ManagedBehaviour
                 Destroy(x.gameObject);
         });
         displayedCards.Clear();
-        int cardsPerRow = Mathf.Max(maxCardsPerRow, cards.Count / numRows);
-        int cardsToPlace = Mathf.Min(maxRows * maxCardsPerRow, cards.Count); // can only show 42 cards max in the selection sequence
-        for (int i = 0; i < cardsToPlace; i++)
+
+        // can only show 42 cards per page
+        int maxPerPage = maxRows * maxCardsPerRow;
+        int startingIndex = maxPerPage * pageIndex;
+        int cardsPerRow = Mathf.Min(maxCardsPerRow, cards.Count / numRows);
+        int cardsToPlace = Mathf.Min(maxPerPage, cards.Count - startingIndex);
+
+        for (int i = startingIndex; i < (startingIndex + cardsToPlace); i++)
         {
-            Vector3 cardPosition = GetCardPosition(i, cardsToPlace, numRows, cardsPerRow);
+            // correct for the current page
+            Vector3 cardPosition = GetCardPosition(i - startingIndex, cardsToPlace, numRows, cardsPerRow);
             PixelPlayableCard card = CreateAndPlaceCard(cards[i], cardPosition);
             if (card == null)
                 yield break;
@@ -171,7 +277,8 @@ public class PixelPlayableCardArray : ManagedBehaviour
             gamepadGrid.Rows = new List<GamepadGridControl.Row>();
         }
         gamepadGrid.Rows.Clear();
-        for (int i = 0; i < maxRows; i++)
+        // add a row for each possible row of cards, plus 1 for the buttons
+        for (int i = 0; i < maxRows + 1; i++)
         {
             gamepadGrid.Rows.Add(new GamepadGridControl.Row());
         }
@@ -214,19 +321,6 @@ public class PixelPlayableCardArray : ManagedBehaviour
         }
     }
 
-    private int GetNumRows(int numCards)
-    {
-        int numRows = 1;
-        int counter = numCards;
-        while (counter > 7)
-        {
-            counter -= maxCardsPerRow;
-            if (counter > 0)
-                numRows++;
-        }
-        return numRows;
-    }
-
     private void SetOverlayEnabled(bool enabled)
     {
         bool swap = false;
@@ -237,14 +331,8 @@ public class PixelPlayableCardArray : ManagedBehaviour
         }
     }
 
-    private void SetCardsEnabled(bool enabled)
-    {
-        foreach (PixelPlayableCard displayedCard in displayedCards)
-        {
-            displayedCard.SetEnabled(enabled);
-        }
-    }
-
+    private int GetNumRows(int numCards) => Mathf.Min(maxRows, numCards / maxCardsPerRow);
+    private void SetCardsEnabled(bool enabled) => displayedCards.ForEach(x => x.SetEnabled(enabled));
     private void SetBoardEnabled(bool enabled)
     {
         IEnumerable<PixelPlayableCard> cards = PixelBoardManager.Instance.CardsOnBoard
@@ -256,11 +344,10 @@ public class PixelPlayableCardArray : ManagedBehaviour
             .Concat(PixelBoardManager.Instance.OpponentQueueSlots)
             .Cast<HighlightedInteractable>();
 
-        HammerButton button = (HammerButton)Resources.FindObjectsOfTypeAll(typeof(HammerButton)).FirstOrDefault();
         PixelCombatBell bell = (PixelCombatBell)Resources.FindObjectsOfTypeAll(typeof(PixelCombatBell)).FirstOrDefault();
 
         bell?.SetEnabled(enabled);
-        button?.gameObject.SetActive(enabled);
+        Singleton<HammerButton>.Instance.gameObject.SetActive(enabled);
 
         foreach (HighlightedInteractable slot in slots)
         {